requirements.txt
python-telegram-bot==20.5
httpx==0.24.1
httpcore==0.17.3
anyio==4.11.0
h11==0.14.0